Remove io features; use only slow io (actually fast)
